# Backend/labfastorder-openapi.yaml
openapi: 3.1.0
info:
  title: "labfastorder API"
  description: "labfastorder API"
  version: "1.0.0"
servers:
  - url: "localhost:8080"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProductCategoryDTO:
      type: "object"
      properties:
        categoryId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        description:
          type: "string"
        productsId:
          type: "array"
          items:
            type: "integer"
            format: "int64"
    NewProductCategoryDTO:
      type: "object"
      properties:
        name:
          type: "string"
        description:
          type: "string"
    OrderDetailDTO:
      type: "object"
      properties:
        orderDetailId:
          type: "integer"
          format: "int64"
        quantity:
          type: "integer"
          format: "int32"
        price:
          type: "number"
          format: "double"
        productId:
          type: "integer"
          format: "int64"
        productName:
          type: "string"
        productSize:
          type: "string"
    Tables:
      type: "object"
      properties:
        tableId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        sizex:
          type: "integer"
          format: "int32"
        sizey:
          type: "integer"
          format: "int32"
        locx:
          type: "integer"
          format: "int32"
        locy:
          type: "integer"
          format: "int32"
    NewTableDTO:
      type: "object"
      properties:
        name:
          type: "string"
        sizex:
          type: "integer"
          format: "int32"
        sizey:
          type: "integer"
          format: "int32"
        locx:
          type: "integer"
          format: "int32"
        locy:
          type: "integer"
          format: "int32"
    ProductDTO:
      type: "object"
      properties:
        productId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        price:
          type: "number"
          format: "double"
        imgName:
          type: "string"
        available:
          type: "boolean"
        quantity:
          type: "integer"
          format: "int32"
        categoryId:
          type: "integer"
          format: "int64"
        size:
          type: "string"
        allergens:
          type: "string"
        ingredients:
          type: "string"
        nutrition:
          type: "string"
    NewProductDTO:
      type: "object"
      properties:
        name:
          type: "string"
        price:
          type: "number"
          format: "double"
        imgName:
          type: "string"
        quantity:
          type: "integer"
          format: "int32"
        productCategoryId:
          type: "integer"
          format: "int64"
        size:
          type: "string"
        allergens:
          type: "string"
        ingredients:
          type: "string"
        nutrition:
          type: "string"
    OrderDTO:
      type: "object"
      properties:
        orderId:
          type: "integer"
          format: "int64"
        status:
          type: "string"
        datetime:
          type: "string"
        totalPrice:
          type: "number"
          format: "double"
        orderDetails:
          type: "array"
          items:
            $ref: "#/components/schemas/OrderDetailDTO"
        tableId:
          type: "integer"
          format: "int64"
...
security:
  - bearerAuth: []
paths:
  /api/productCategories:
    get:
      summary: "GET api/productCategories"
      operationId: "getAllProductCategories"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ProductCategoryDTO"
    post:
      summary: "POST api/productCategories"
      operationId: "createProductCategory"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewProductCategoryDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductCategoryDTO"
  /api/productCategories/{id}:
    get:
      summary: "GET api/productCategories/{id}"
      operationId: "getProductCategoryById"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductCategoryDTO"
    put:
      summary: "PUT api/productCategories/{id}"
      operationId: "updateProductCategory"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewProductCategoryDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductCategoryDTO"
    delete:
      summary: "DELETE api/productCategories/{id}"
      operationId: "deleteProductCategory"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...
  /api/tables:
    get:
      summary: "GET api/tables"
      operationId: "getAllTables"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Tables"
    post:
      summary: "POST api/tables"
      operationId: "createTable"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTableDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Tables"
  /api/tables/{id}:
    get:
      summary: "GET api/tables/{id}"
      operationId: "getTableById"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Tables"
    put:
      summary: "PUT api/tables/{id}"
      operationId: "updateTable"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTableDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Tables"
    delete:
      summary: "DELETE api/tables/{id}"
      operationId: "deleteTables"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...
  /api/products/{id}:
    get:
      summary: "GET api/products/{id}"
      operationId: "getProductById"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      summary: "PUT api/products/{id}"
      operationId: "updateProduct"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewProductDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      summary: "DELETE api/products/{id}"
      operationId: "deleteProduct"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...
  /api/orders/{id}:
    get:
      summary: "GET api/orders/{id}"
      operationId: "getOrderById"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    delete:
      summary: "DELETE api/orders/{id}"
      operationId: "deleteOrder"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...